third_party/blink/web_tests/external/wpt/css/css-values/signs-abs-serialize-expected.txt

This is a testharness.js-based test.
Found 56 tests; 4 PASS, 52 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL 'abs(1)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'abs(1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(abs(1))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(abs(1))' should be valid in transform. got disallowed value ""
PASS 'abs(1)' as a computed value should serialize as '1'.
FAIL 'scale(abs(1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(abs(1))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'sign(.1)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'sign(.1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(sign(.1))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(sign(.1))' should be valid in transform. got disallowed value ""
PASS 'sign(.1)' as a computed value should serialize as '1'.
FAIL 'scale(sign(.1))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(sign(.1))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'abs(1 + 2 + 3)' as a specified value should serialize as 'calc(6)'. assert_not_equals: 'abs(1 + 2 + 3)' should be valid in opacity. got disallowed value ""
FAIL 'scale(abs(1 + 2 + 3))' as a specified value should serialize as 'scale(calc(6))'. assert_not_equals: 'scale(abs(1 + 2 + 3))' should be valid in transform. got disallowed value ""
FAIL 'abs(1 + 2 + 3)' as a computed value should serialize as '6'. assert_equals: '6' should round-trip exactly in computed values. expected "6" but got "1"
FAIL 'scale(abs(1 + 2 + 3))' as a computed value should serialize as 'matrix(6, 0, 0, 6, 0, 0)'. assert_equals: 'scale(abs(1 + 2 + 3))' and 'matrix(6, 0, 0, 6, 0, 0)' should serialize the same in computed values. expected "matrix(6, 0, 0, 6, 0, 0)" but got "none"
FAIL 'sign(1 + 2 + 3)' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'sign(1 + 2 + 3)' should be valid in opacity. got disallowed value ""
FAIL 'scale(sign(1 + 2 + 3))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(sign(1 + 2 + 3))' should be valid in transform. got disallowed value ""
PASS 'sign(1 + 2 + 3)' as a computed value should serialize as '1'.
FAIL 'scale(sign(1 + 2 + 3))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(sign(1 + 2 + 3))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'calc(abs(1) + abs(2))' as a specified value should serialize as 'calc(3)'. assert_not_equals: 'calc(abs(1) + abs(2))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(abs(1) + abs(2)))' as a specified value should serialize as 'scale(calc(3))'. assert_not_equals: 'scale(calc(abs(1) + abs(2)))' should be valid in transform. got disallowed value ""
FAIL 'calc(abs(1) + abs(2))' as a computed value should serialize as '3'. assert_equals: '3' should round-trip exactly in computed values. expected "3" but got "1"
FAIL 'scale(calc(abs(1) + abs(2)))' as a computed value should serialize as 'matrix(3, 0, 0, 3, 0, 0)'. assert_equals: 'scale(calc(abs(1) + abs(2)))' and 'matrix(3, 0, 0, 3, 0, 0)' should serialize the same in computed values. expected "matrix(3, 0, 0, 3, 0, 0)" but got "none"
FAIL 'calc(sign(.1) + sign(.2))' as a specified value should serialize as 'calc(2)'. assert_not_equals: 'calc(sign(.1) + sign(.2))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sign(.1) + sign(.2)))' as a specified value should serialize as 'scale(calc(2))'. assert_not_equals: 'scale(calc(sign(.1) + sign(.2)))' should be valid in transform. got disallowed value ""
FAIL 'calc(sign(.1) + sign(.2))' as a computed value should serialize as '2'. assert_equals: '2' should round-trip exactly in computed values. expected "2" but got "1"
FAIL 'scale(calc(sign(.1) + sign(.2)))' as a computed value should serialize as 'matrix(2, 0, 0, 2, 0, 0)'. assert_equals: 'scale(calc(sign(.1) + sign(.2)))' and 'matrix(2, 0, 0, 2, 0, 0)' should serialize the same in computed values. expected "matrix(2, 0, 0, 2, 0, 0)" but got "none"
FAIL 'calc(1 + abs(1))' as a specified value should serialize as 'calc(2)'. assert_not_equals: 'calc(1 + abs(1))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(1 + abs(1)))' as a specified value should serialize as 'scale(calc(2))'. assert_not_equals: 'scale(calc(1 + abs(1)))' should be valid in transform. got disallowed value ""
FAIL 'calc(1 + abs(1))' as a computed value should serialize as '2'. assert_equals: '2' should round-trip exactly in computed values. expected "2" but got "1"
FAIL 'scale(calc(1 + abs(1)))' as a computed value should serialize as 'matrix(2, 0, 0, 2, 0, 0)'. assert_equals: 'scale(calc(1 + abs(1)))' and 'matrix(2, 0, 0, 2, 0, 0)' should serialize the same in computed values. expected "matrix(2, 0, 0, 2, 0, 0)" but got "none"
FAIL 'calc(sign(.1) + 1)' as a specified value should serialize as 'calc(2)'. assert_not_equals: 'calc(sign(.1) + 1)' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sign(.1) + 1))' as a specified value should serialize as 'scale(calc(2))'. assert_not_equals: 'scale(calc(sign(.1) + 1))' should be valid in transform. got disallowed value ""
FAIL 'calc(sign(.1) + 1)' as a computed value should serialize as '2'. assert_equals: '2' should round-trip exactly in computed values. expected "2" but got "1"
FAIL 'scale(calc(sign(.1) + 1))' as a computed value should serialize as 'matrix(2, 0, 0, 2, 0, 0)'. assert_equals: 'scale(calc(sign(.1) + 1))' and 'matrix(2, 0, 0, 2, 0, 0)' should serialize the same in computed values. expected "matrix(2, 0, 0, 2, 0, 0)" but got "none"
FAIL 'calc(abs(inFinity))' as a specified value should serialize as 'calc(inFinity)'. assert_not_equals: 'calc(abs(inFinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(abs(inFinity)))' as a specified value should serialize as 'scale(calc(infinity))'. assert_not_equals: 'scale(calc(abs(inFinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(abs(inFinity))' as a computed value should serialize as 'infinity'. assert_equals: 'infinity' should round-trip exactly in computed values. expected "infinity" but got "1"
FAIL 'scale(calc(abs(inFinity)))' as a computed value should serialize as 'matrix(infinity, 0, 0, infinity, 0, 0)'. assert_equals: 'matrix(infinity, 0, 0, infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(infinity, 0, 0, infinity, 0, 0)" but got "none"
FAIL 'calc(abs(infinity))' as a specified value should serialize as 'calc(infinity)'. assert_not_equals: 'calc(abs(infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(abs(infinity)))' as a specified value should serialize as 'scale(calc(infinity))'. assert_not_equals: 'scale(calc(abs(infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(abs(infinity))' as a computed value should serialize as 'infinity'. assert_equals: 'infinity' should round-trip exactly in computed values. expected "infinity" but got "1"
FAIL 'scale(calc(abs(infinity)))' as a computed value should serialize as 'matrix(infinity, 0, 0, infinity, 0, 0)'. assert_equals: 'matrix(infinity, 0, 0, infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(infinity, 0, 0, infinity, 0, 0)" but got "none"
FAIL 'calc(sign(infinity))' as a specified value should serialize as 'calc(1)'. assert_not_equals: 'calc(sign(infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sign(infinity)))' as a specified value should serialize as 'scale(calc(1))'. assert_not_equals: 'scale(calc(sign(infinity)))' should be valid in transform. got disallowed value ""
PASS 'calc(sign(infinity))' as a computed value should serialize as '1'.
FAIL 'scale(calc(sign(infinity)))' as a computed value should serialize as 'matrix(1, 0, 0, 1, 0, 0)'. assert_equals: 'scale(calc(sign(infinity)))' and 'matrix(1, 0, 0, 1, 0, 0)' should serialize the same in computed values. expected "matrix(1, 0, 0, 1, 0, 0)" but got "none"
FAIL 'abs(infinity)' as a specified value should serialize as 'calc(infinity)'. assert_not_equals: 'abs(infinity)' should be valid in opacity. got disallowed value ""
FAIL 'scale(abs(infinity))' as a specified value should serialize as 'scale(calc(infinity))'. assert_not_equals: 'scale(abs(infinity))' should be valid in transform. got disallowed value ""
FAIL 'abs(infinity)' as a computed value should serialize as 'infinity'. assert_equals: 'infinity' should round-trip exactly in computed values. expected "infinity" but got "1"
FAIL 'scale(abs(infinity))' as a computed value should serialize as 'matrix(infinity, 0, 0, infinity, 0, 0)'. assert_equals: 'matrix(infinity, 0, 0, infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(infinity, 0, 0, infinity, 0, 0)" but got "none"
FAIL 'calc(abs(-infinity))' as a specified value should serialize as 'calc(infinity)'. assert_not_equals: 'calc(abs(-infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(abs(-infinity)))' as a specified value should serialize as 'scale(calc(infinity))'. assert_not_equals: 'scale(calc(abs(-infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(abs(-infinity))' as a computed value should serialize as 'infinity'. assert_equals: 'infinity' should round-trip exactly in computed values. expected "infinity" but got "1"
FAIL 'scale(calc(abs(-infinity)))' as a computed value should serialize as 'matrix(infinity, 0, 0, infinity, 0, 0)'. assert_equals: 'matrix(infinity, 0, 0, infinity, 0, 0)' should round-trip exactly in computed values. expected "matrix(infinity, 0, 0, infinity, 0, 0)" but got "none"
FAIL 'calc(sign(-1 * infinity))' as a specified value should serialize as 'calc(-1)'. assert_not_equals: 'calc(sign(-1 * infinity))' should be valid in opacity. got disallowed value ""
FAIL 'scale(calc(sign(-1 * infinity)))' as a specified value should serialize as 'scale(calc(-1))'. assert_not_equals: 'scale(calc(sign(-1 * infinity)))' should be valid in transform. got disallowed value ""
FAIL 'calc(sign(-1 * infinity))' as a computed value should serialize as '-1'. assert_equals: '-1' should round-trip exactly in computed values. expected "-1" but got "0"
FAIL 'scale(calc(sign(-1 * infinity)))' as a computed value should serialize as 'matrix(-1, 0, 0, -1, 0, 0)'. assert_equals: 'scale(calc(sign(-1 * infinity)))' and 'matrix(-1, 0, 0, -1, 0, 0)' should serialize the same in computed values. expected "matrix(-1, 0, 0, -1, 0, 0)" but got "none"
Harness: the test ran to completion.
